ir-service/requirements.txt
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
PyYAML==3.12
uWSGI==2.0.14
Werkzeug==0.11.15
